fix: WakuMessage json encoding
